Allow specifying the allowed IPs in a message
This lets you do config like:
 $wgNetworkAuthUsers[] = [
	'ipmsg' => 'networkauth-ips',
	'user' => 'Foo'
 );

And then have the list of IPs be on MediaWiki:networkauth-ips

The format of the message is a * followed by either an
IP address (192.168.0.1), a plain range (10.0.0.3 - 10.0.0.5)
or a CIDR range (10.1.0.0/16). Any line not not starting with a *
is considered a comment and ignored.

Change-Id: I504258a41208a298a989a137b7da92f9f027ce2e